added GPU check and did some GPU refactoring.

+++ motif_tools.py
@@ -190,7 +190,7 @@ class Motif():
 
     def build_ddg_array(self):
         ref_energy = self.consensus_energy
-        energies = numpy.zeros(3*len(self), dtype=float)
+        energies = numpy.zeros(3*len(self), dtype='float32')
         for i, base_energies in enumerate(self.motif_data):
             for j, base_energy in enumerate(base_energies[1:]):
                 energies[3*i+j] = base_energy - base_energies[0]
